Karma to Jest: team_details_ctrl (#12321)
